% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_GATE.R
\name{aipw_GATE}
\alias{aipw_GATE}
\title{Pseudo-outcomes AIPW estimation}
\usage{
aipw_GATE(X, W, Y, e, nfolds)
}
\arguments{
\item{X}{Feature matrix}

\item{W}{Binary treatment indicator}

\item{Y}{Response variable vector}

\item{e}{Propensity score: true or estimated}

\item{nfolds}{Numbers of folds for cross-fitting}
}
\value{
Returns a vector of pseudo-outcomes
}
\description{
Pseudo-outcomes AIPW estimation
}
\examples{
Not yet
}
