% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_learner.R
\name{DR_learner}
\alias{DR_learner}
\title{DR-learner}
\usage{
DR_learner(X, W, Y, train_idx, e)
}
\arguments{
\item{X}{Feature matrix}

\item{W}{Binary treatment indicator}

\item{Y}{Response variable vector}

\item{train_idx}{Training sample indicator:
"1" for train,
"0" for test}

\item{e}{Propensity score: true or estimated}
}
\value{
Returns CATE values estimated on the "train" sample and predicted on the "test" sample
}
\description{
Estimates and Predicts DR-learner using regression forests.
The estimation is conducted in full-sample.
}
\examples{
Not yet
}
