% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{monte_carlo}
\alias{monte_carlo}
\title{Monte Carlo draws for coverage}
\usage{
monte_carlo(X, W, Y, e, iterations, train_share)
}
\arguments{
\item{X}{Feature matrix}

\item{W}{Binary treatment indicator}

\item{Y}{Response variable vector}

\item{e}{Propensity score: true or estimated}

\item{iterations}{Number of times to rerun the analysis}

\item{train_share}{Share of the training sample relative to the full data}
}
\value{
Returns shares of confidence intervals of estimated HTE b2 with does NOT cover zero.
}
\description{
This function is a bit challenging. It also takes a long time to run. I use it to analyse how the results depend on seed number as well as train/test shares.
}
\examples{
Not yet

}
