% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_learner.R
\name{R_learner}
\alias{R_learner}
\title{R-learner}
\usage{
R_learner(X, W, Y, train_idx, e)
}
\arguments{
\item{X}{Feature matrix}

\item{W}{Binary treatment indicator}

\item{Y}{Response variable vector}

\item{train_idx}{Training sample indicator:
"1" for train,
"0" for test}

\item{e}{Propensity score: true or estimated}
}
\value{
Returns CATE values estimated on the "train" sample and predicted on the "test" sample
}
\description{
Estimates and Predicts R-learner using regression forests.
The estimation is conducted in full-sample.
}
\examples{
Not yet
}
