% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_learner.R
\name{T_learner}
\alias{T_learner}
\title{T-learner}
\usage{
T_learner(X, W, Y, train_idx)
}
\arguments{
\item{X}{Feature matrix}

\item{W}{Binary treatment indicator}

\item{Y}{Response variable vector}

\item{train_idx}{Training sample indicator:
"1" for train,
"0" for test}
}
\value{
Returns CATE values estimated on the "train" sample and predicted on the "test" sample
}
\description{
Estimates and Predicts T-learner using regression forests.
The estimation is conducted in full-sample.
}
\examples{
Not yet
}
